[SPARK-20231] Refactor star schema code.
